+++ packages/db/src/index.ts
@@ -1,0 +1,2 @@
+export { connectDB, mongoose } from "./connection";
+export * from "./models";